+++ src/Template/Form/Field.php
@@ -16,10 +16,19 @@ final class Field implements NameTypeItem, ControlInterface, JsonSerializable
 
     private Type $type;
 
-    public function __construct(string $name, Type $type)
+    /**
+     * @var ?array<int|string, int|string>
+     */
+    private ?array $options;
+
+    /**
+     * @param ?array<int|string, int|string> $options - we don't care about values, so we use only keys as keys and also as values here
+     */
+    public function __construct(string $name, Type $type, ?array $options = null)
     {
         $this->name = $name;
         $this->type = $type;
+        $this->options = $options;
     }
 
     public function getName(): string
@@ -44,12 +53,21 @@ final class Field implements NameTypeItem, ControlInterface, JsonSerializable
         return $clone;
     }
 
+    /**
+     * @return ?array<int|string, int|string>
+     */
+    public function getOptions(): ?array
+    {
+        return $this->options;
+    }
+
     #[ReturnTypeWillChange]
     public function jsonSerialize()
     {
         return [
-          'name' => $this->name,
-          'type' => $this->getTypeAsString(),
+            'name' => $this->name,
+            'type' => $this->getTypeAsString(),
+            'options' => $this->options,
         ];
     }
 }